fix types for import/export space
